snackbar showing points per day exceeded

+++ app/src/main/java/com/blogspot/android_czy_java/apps/mgr/main/chat/usecase/ObserveFirestoreMessages.kt
@@ -59,8 +59,8 @@ class ObserveFirestoreMessages @Inject constructor(
                 messagesToAdd.filter {
                     it.authorId == FirebaseAuth.getInstance().currentUser?.uid
                 }
-                    .count()
-            if (points != 0) {
+                    .count().toLong()
+            if (points != 0L) {
                 coursesDao.setPoints(messagesToAdd[0].courseId, points)
             }
         }
