+++ app/src/main/java/com/blogspot/android_czy_java/apps/mgr/main/course/CourseFragmentPresenter.kt
@@ -15,7 +15,7 @@ class CourseFragmentPresenter @Inject constructor(private val courseDao: Courses
     fun getCourseTitle(): LiveData<String> =
         courseDao.getCourseTitle(courseId)
 
-    fun getActivityPoints(): LiveData<Int> =
+    fun getActivityPoints(): LiveData<Long> =
         courseDao.getPoints(courseId)
